don't use require after downloading solidity
This is a fix so that remix will work inside of electron, and presumably, it makes sense for dynamic downloading of solidity compiler. When I enable node integration with electron and install solc node module, it looks like it works for a second, but then it too does not work (page goes blank).

I don't think require() works here, because for that it work with node, it would need to be downloaded into a path that can be resolved by require (typically node_modules). Notably, electron docs recommend against enabling node integration when remote scripts are loaded dynamically.